agentic workflow app/requirements.txt
streamlit
openai
python-dotenv
selenium
beautifulsoup4
webdriver-manager